@@ -12,22 +12,24 @@
     ]
   },
   "packageManager": "yarn@1.22.22",
+  "engines": {
+    "node": ">=20.18.0"
+  },
   "dependencies": {
     "axios": "^0.21.1",
     "date-fns": "^2.16.1",
     "date-fns-tz": "^1.1.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^4.1.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.4"
+    "react-router-dom": "^6.28.0",
+    "react-scripts": "^5.0.1"
   },
   "devDependencies": {
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.38",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router-dom": "^5.1.7",
-    "typescript": "^4.1.3"
+    "@types/node": "^20.17.10",
+    "@types/react": "^18.3.16",
+    "@types/react-dom": "^18.3.5",
+    "typescript": "^5.7.2"
   },
   "name": "pscj-web",
   "private": true,
