@@ -1,15 +1,15 @@
-import React, { useState, FormEvent, useEffect } from 'react'
-import { useHistory } from 'react-router-dom'
 import { format } from 'date-fns'
+import React, { FormEvent, useEffect, useState } from 'react'
 import { BiChurch } from 'react-icons/bi'
 import { FaMapMarkedAlt } from 'react-icons/fa'
 import { HiUserGroup } from 'react-icons/hi'
+import { useNavigate } from 'react-router-dom'
 
-import api from '../../services/api'
-import { Missa, Local } from '../../utils/interfaces'
-import { dataParaSelect } from '../../utils/tratandoDatas'
 import flechaTorta from '../../assets/icons/flechaTorta.png'
 import sublinhado from '../../assets/sublinhado.png'
+import { Local, Missa } from '../../utils/interfaces'
+import { mockedLocais } from '../../utils/mocks/locais'
+import { dataParaSelect } from '../../utils/tratandoDatas'
 
 import './styles.css'
 
@@ -28,47 +28,59 @@ const FormMissas: React.FC<FormMissa> = ({ titulo, txtBtn, missa, mensagemEsquer
   const [max_pessoas, setMax_pessoas] = useState(missa?.max_pessoas || '')
   const [data_hora, setData_hora] = useState(missa?.data_hora ? dataParaSelect(missa.data_hora) : '')
 
-  const { push } = useHistory()
+  const navigate = useNavigate()
 
   useEffect(() => {
-    api.get('locais')
-      .then(({ data }) => setLocais(data))
-      .catch(({ response }) => {
-        console.log(response)
-        alert(response?.data.erro || 'Falha ao listar locais.')
-      })
+    // api.get('locais')
+    //   .then(({ data }) => setLocais(data))
+    //   .catch(({ response }) => {
+    //     console.log(response)
+    //     alert(response?.data.erro || 'Falha ao listar locais.')
+    //   })
+
+    setLocais(mockedLocais)
   }, [])
 
   function criarEditarMissa(event: FormEvent) {
     event.preventDefault()
 
     if (data_hora) {
-      const dadosMissa = { nome, local_id, data_hora, max_pessoas }
+      // const dadosMissa = { nome, local_id, data_hora, max_pessoas }
 
       if (!missa) {
-        api.post('missas', dadosMissa)
-          .then(({ data }) => {
-            setNome('')
-            setLocal_id(0)
-            setMax_pessoas('')
-            setData_hora('')
+        // api.post('missas', dadosMissa)
+        //   .then(({ data }) => {
+        //     setNome('')
+        //     setLocal_id(0)
+        //     setMax_pessoas('')
+        //     setData_hora('')
 
-            alert(data.mensagem)
-          })
-          .catch(({ response }) => {
-            console.log(response)
-            alert(response.data.erro || 'Falha ao cadastrar a missa')
-          })
+        //     alert(data.mensagem)
+        //   })
+        //   .catch(({ response }) => {
+        //     console.log(response)
+        //     alert(response.data.erro || 'Falha ao cadastrar a missa')
+        //   })
+
+        setNome('')
+        setLocal_id(0)
+        setMax_pessoas('')
+        setData_hora('')
+
+        alert('Missa cadastrada com sucesso!')
       } else {
-        api.put(`missas/${missa.id}`, dadosMissa)
-          .then(({ data }) => {
-            alert(data.mensagem)
-            push('/lista-missas')
-          })
-          .catch(({ response }) => {
-            console.log(response)
-            alert(response.data.erro || 'Falha ao atualizar a missa')
-          })
+        // api.put(`missas/${missa.id}`, dadosMissa)
+        //   .then(({ data }) => {
+        //     alert(data.mensagem)
+        //     push('/lista-missas')
+        //   })
+        //   .catch(({ response }) => {
+        //     console.log(response)
+        //     alert(response.data.erro || 'Falha ao atualizar a missa')
+        //   })
+
+        alert('Missa atualizada com sucesso!')
+        navigate('/lista-missas')
       }
     }
   }
