@@ -1,8 +1,8 @@
-import React from 'react'
-import { FaFacebookF, FaInstagram, FaYoutube, FaWhatsapp } from 'react-icons/fa'
+import { FaFacebookF, FaInstagram, FaWhatsapp, FaYoutube } from 'react-icons/fa'
 import { Link } from 'react-router-dom'
 
 import flechaTorta from '../../assets/icons/flechaTorta.png'
+
 import './styles.css'
 
 const Footer = () => {
