@@ -1,5 +1,5 @@
-import React from 'react'
-import { render } from 'react-dom'
+import { createRoot } from 'react-dom/client'
+
 import App from './App'
 
-render(<App />, document.getElementById('root'))
+createRoot(document.getElementById('root')!).render(<App />)
